fixed setAngle signal to type double. Connected signal and slot. Works fine.

+++ kinect/src/gui.h
@@ -61,7 +61,7 @@ class Gui : public QMainWindow
 		void setRedLedFlash();
 		void setGreenLedFlash();
 		void setYellowLedFlash();
-		void setAngle(int i);
+		void setAngle(double i);
 
 
 	protected:
